🚚 Move: components for top page to `$lib/components/home/`
