+++ src/lib/components/home/Activities.svelte
