+++ src/lib/components/home/Chronicle.svelte
