+++ src/lib/components/home/MinimalContactTable.svelte
